+++ test/test_file_util.py
@@ -4,6 +4,7 @@ from eloservice.file_util import FileUtil
 from eloservice.login_util import LoginUtil
 from test import TEST_ROOT_DIR, _check_not_unset
 import os
+import tempfile
 
 
 class TestService(unittest.TestCase):
@@ -63,11 +64,35 @@ class TestService(unittest.TestCase):
         elo_connection, elo_client = self._login()
         util = FileUtil(elo_client, elo_connection)
         parentID = "134698"
-        sord = util.upload_file(TEST_ROOT_DIR + "/resources/testFile.png", parentID, filename="FileNameTestFile.png",
-                                filename_objkey="differentFilename.png")
+        sord = util.upload_file(TEST_ROOT_DIR + "/resources/chicken.pdf", parentID, filename="importantDokument",
+                                filename_objkey="chicken.pdf")
         assert sord is not None
         assert sord != ""
         checkSordID = sord
         sord = util.checkout_sord(checkSordID)
-        assert sord.name == "FileNameTestFile.png"
-        assert sord.obj_keys[0].data[0] == "differentFilename.png"
+        assert sord.name == "importantDokument"
+        assert sord.obj_keys[0].data[0] == "chicken.pdf"
+
+    def test_upload_file_with_custom_filedate(self):
+        elo_connection, elo_client = self._login()
+        util = FileUtil(elo_client, elo_connection)
+        parentID = "134698"
+        sord = util.upload_file(TEST_ROOT_DIR + "/resources/chicken.pdf", parentID, filename="customFiledate",
+                                filename_objkey="chicken.pdf",
+                                filedate="2021-08-25T15:00:00")
+        assert sord is not None
+        assert sord != ""
+        checkSordID = sord
+        sord = util.checkout_sord(checkSordID)
+        assert sord.name == "customFiledate"
+        assert sord.x_date_iso == "20210825150000"
+        assert sord.obj_keys[0].data[0] == "chicken.pdf"
+
+    def test_download_file(self):
+        elo_connection, elo_client = self._login()
+        util = FileUtil(elo_client, elo_connection)
+        sordID = "143926"
+        temp_file_path = tempfile.mktemp()
+        file = util.download_file(sordID, temp_file_path)
+        assert file is not None
+        assert file != ""